requirements.txt
streamlit==1.32.0
pandas 
plotly 
mysql-connector-python 
python-dotenv
crewai==0.11.0
langchain 
langchain-community
